sna: Clean up some debug messages for 64bit

+++ src/sna/sna_dri.c
@@ -630,7 +630,7 @@ sna_dri_frame_event_drawable_gone(void *data, XID id)
 					 drawable_resource);
 
 		DBG(("%s: marking drawable gone [%p]: %ld\n",
-		     __FUNCTION__, info, info->drawable_id));
+		     __FUNCTION__, info, (long)info->drawable_id));
 
 		list_del(&info->drawable_resource);
 		info->drawable_id = None;
@@ -696,7 +696,7 @@ sna_dri_add_frame_event(struct sna_dri_frame_event *info)
 	list_add(&info->drawable_resource, resource);
 
 	DBG(("%s: add[%p] (%p, %ld)\n", __FUNCTION__,
-	     info, info->client, info->drawable_id));
+	     info, info->client, (long)info->drawable_id));
 
 	return TRUE;
 }
@@ -705,7 +705,7 @@ static void
 sna_dri_frame_event_info_free(struct sna_dri_frame_event *info)
 {
 	DBG(("%s: del[%p] (%p, %ld)\n", __FUNCTION__,
-	     info, info->client, info->drawable_id));
+	     info, info->client, (long)info->drawable_id));
 
 	list_del(&info->client_resource);
 	list_del(&info->drawable_resource);
